Added useStopwatch hook
