+++ src/components/StopWatch.tsx
@@ -3,52 +3,19 @@ import StopWatchButton from './StopWatchButton';
 import LapsList from './LapsList';
 import Time from './Time';
 import * as mq from '../styles/media-queries';
-
-// Function to get stored time from local storage,
-function setStoredTime() {
-  const storedTime = localStorage.getItem('stopwatch_time');
-  return storedTime ? parseFloat(storedTime) : 0;
-}
-
-// Function to get stored laps from local storage
-function setStoredLaps() {
-  const storedLaps = localStorage.getItem('stopwatch_laps');
-  return storedLaps ? JSON.parse(storedLaps) : [];
-}
+import useStopwatch from '../hooks/useStopwatch';
+import { css } from '@emotion/react';
 
 export default function StopWatch() {
-  const [time, setTime] = useState(setStoredTime);
-  const [laps, setLaps] = useState(setStoredLaps);
-  const [isStarted, setIsStarted] = useState(false);
-  const requestIdRef = useRef<number>();
-
-  // Main 
-  useEffect(() => {
-    let prevTime: number;
-
-    if (isStarted) {
-      requestIdRef.current = requestAnimationFrame(function updateTimer(
-        timestamp: number
-      ) {
-        if (!prevTime) prevTime = timestamp;
-        const deltaMs = timestamp - prevTime;
-        prevTime = timestamp;
-        setTime((prev => prev + deltaMs));
-        requestIdRef.current = requestAnimationFrame(updateTimer);
-      });
-    } else {
-      cancelAnimationFrame(requestIdRef.current);
-    }
-    // Clean up
-    return () => {
-      cancelAnimationFrame(requestIdRef.current);
-    };
-  }, [isStarted]);
-
-  useEffect(() => {
-    // Save laps to local storage
-    localStorage.setItem('stopwatch_laps', JSON.stringify(laps));
-  }, [laps]);
+  const {
+    time,
+    laps,
+    isStarted,
+    startTimer,
+    stopTimer,
+    resetTimer,
+    addLap,
+  } = useStopwatch();
 
   useEffect(() => {
     if (isStarted) {
@@ -60,97 +27,59 @@ export default function StopWatch() {
     window.removeEventListener('beforeunload', handleBeforeUnload);
   }, [isStarted]);
 
-  const startTimerHandler = () => {
-    setIsStarted(true);
-  };
-
-  const stopTimerHandler = () => {
-    setIsStarted(false);
-    // Save time to local storage
-    localStorage.setItem('stopwatch_time', time.toString());
-  };
-
-  const resetTimerHandler = () => {
-    setIsStarted(false);
-    setTime(0);
-    setLaps([]);
-    // Reset local storage
-    localStorage.setItem('stopwatch_time', '0');
-    localStorage.setItem('stopwatch_laps', JSON.stringify([]));
-  };
-
-  const addLap = () => {
-    // Get the last lap
-    const lastLap = laps.length > 0 ? laps[laps.length - 1] : null;
-
-    // Check if there is a last lap and if the current time is different from the last lap's timestamp
-    if (lastLap) {
-      if (lastLap.timestamp === time) return;
-      setLaps((prevLaps: { timestamp: number; duration: number }[]) => [
-        ...prevLaps,
-        { timestamp: time, duration: time - lastLap.timestamp },
-      ]);
-    } else {
-      setLaps((prevLaps: { timestamp: number; duration: number }[]) => [
-        ...prevLaps,
-        { timestamp: time, duration: time },
-      ]);
-    }
-  };
-
   //Prevents immediate refreshing/navigating when stopwatch is running
   const handleBeforeUnload = (event: BeforeUnloadEvent) => {
     event.preventDefault();
     event.returnValue = 'Are you sure you want to exit?';
     const shouldReload = window.confirm;
     if (shouldReload) {
-      resetTimerHandler()
+      resetTimer()
     }
     return event.returnValue;
   };
 
+  const mainContainerStyle = css ({
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    width: '90%',
+    [mq.large]: {
+      flexDirection: 'row',
+    },
+  })
+
+  const timerContainerStyle =  css({
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'start',
+    minWidth: '100%',
+    [mq.large]: {
+      minWidth: '50%',
+    },
+  })
+
+  const buttonsContainerStyle = css({
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'center',
+    width: '40%',
+    margin: '0 auto',
+    [mq.large]: {
+      flexDirection: 'column',
+    },
+  })
+
   return (
-    <div
-      css={{
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'center',
-        width: '90%',
-        [mq.large]: {
-          flexDirection: 'row',
-        },
-      }}
-    >
-      <div
-        css={{
-          display: 'flex',
-          flexDirection: 'column',
-          justifyContent: 'start',
-          minWidth: '100%',
-          [mq.large]: {
-            minWidth: '50%',
-          },
-        }}
-      >
+    <div css={mainContainerStyle}>
+      <div css={timerContainerStyle}>
         <div css={{ marginBottom: '5%' }} data-testid="timer-container">
           <Time time={time} />
         </div>
 
-        <div
-          css={{
-            display: 'flex',
-            flexDirection: 'row',
-            justifyContent: 'center',
-            width: '40%',
-            margin: '0 auto',
-            [mq.large]: {
-              flexDirection: 'column',
-            },
-          }}
-        >
+        <div css={buttonsContainerStyle}>
           <StopWatchButton
             variant="default"
-            onClick={startTimerHandler}
+            onClick={startTimer}
             disabled={isStarted}
             testId="start-button"
           >
@@ -158,7 +87,7 @@ export default function StopWatch() {
           </StopWatchButton>
           <StopWatchButton
             variant="default"
-            onClick={stopTimerHandler}
+            onClick={stopTimer}
             disabled={!isStarted}
             testId="pause-button"
           >
@@ -166,7 +95,7 @@ export default function StopWatch() {
           </StopWatchButton>
           <StopWatchButton
             variant="destructive"
-            onClick={resetTimerHandler}
+            onClick={resetTimer}
             disabled={time === 0}
             testId="reset-button"
           >
